Reduced number of dependencies to simplify imports
