+++ packages/moocfi-quizzes/example/package.json
@@ -8,23 +8,7 @@
     "build": "parcel build index.html"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-free": "^5.9.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.19",
-    "@fortawesome/free-solid-svg-icons": "^5.9.0",
-    "@fortawesome/react-fontawesome": "^0.1.4",
-    "@material-ui/styles": "^4.2.1",
-    "@types/commonmark": "^0.27.4",
-    "axios": "^0.18.0",
-    "commonmark": "^0.29.0",
-    "likert-react": "^0.0.0-beta6",
-    "react": ">= 16.3.0",
-    "react-app-polyfill": "^1.0.1",
-    "react-dom": ">= 16.3.0",
-    "react-redux": "^7.1.0",
-    "redux": "^4.0.4",
-    "redux-thunk": "^2.3.0",
-    "styled-components": "^4.3.2",
-    "typesafe-actions": "^4.4.2"
+    "react-app-polyfill": "^1.0.1"
   },
   "devDependencies": {
     "@types/react": "^16.8.15",
@@ -32,11 +16,17 @@
     "parcel": "^1.12.3",
     "typescript": "^3.4.5"
   },
-  "alias": {
-    "react": "../node_modules/react",
-    "react-dom": "../node_modules/react-dom",
-    "@material-ui/core": "../node_modules/@material-ui/core",
-    "@material-ui/styles": "../node_modules/@material-ui/styles",
-    "styled-components": "../node_modules/styled-components"
+  "peerDependencies": {
+    "@material-ui/core": ">= 4.0.0",
+    "@material-ui/styles": ">=4.0.0",
+    "commonmark": ">= 0.29.0",
+    "react": ">= 16.8.0",
+    "react-dom": ">= 16.8.0",
+    "react-redux": ">= 7.1.0",
+    "redux": ">= 4.0.4",
+    "redux-thunk": ">= 2.3.0",
+    "styled-components": ">= 4.0.0",
+    "typesafe-actions": ">= 4.4.2"
+
   }
 }
