+++ packages/moocfi-quizzes/example/src/SimpleErrorBoundary.tsx
@@ -1,4 +1,4 @@
-import * as React from "../node_modules/react"
+import * as React from "react"
 
 class SimpleErrorBoundary extends React.Component<any, any> {
   state = {
