+++ packages/moocfi-quizzes/example/src/customHooks.tsx
@@ -1,0 +1,24 @@
+import { useState } from "react"
+
+export function useInputWithLocalStorage(key, initialValue) {
+  const [value, setValue] = useLocalStorage(key, initialValue)
+
+  const onChange = e => {
+    setValue(e.target.value)
+  }
+
+  return { value, onChange }
+}
+
+function useLocalStorage(key, initialValue) {
+  const [storedValue, setStoredValue] = useState(() => {
+    return window.localStorage.getItem(key) || initialValue
+  })
+
+  const setValue = value => {
+    setStoredValue(value)
+    window.localStorage.setItem(key, value)
+  }
+
+  return [storedValue, setValue]
+}
